requirements.txt
asgiref==3.5.2
Django~=4.0
sqlparse==0.4.2
Pillow==9.2.0
flower==1.1.0
celery==5.2.7
stripe==3.5.0
WeasyPrint==56.0
redis==4.3.4
django-rosetta==0.9.8
django-parler==2.3
django-localflavor==3.1
psycopg2==2.9.3
python-decouple==3.6

